Add all CRUD operations
