+++ app/assets/stylesheets/products.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the Product controller here.
+// Place all the styles related to the Products controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: https://sass-lang.com/
