--- app/helpers/product_helper.rb
@@ -1,2 +1,0 @@
-module ProductHelper
-end
