--- app/views/product/delete.html.haml
@@ -1,2 +1,0 @@
-%h1 Product#delete
-%p Find me in app/views/product/delete.html.haml
