--- app/views/product/edit.html.haml
@@ -1,2 +1,0 @@
-%h1 Product#edit
-%p Find me in app/views/product/edit.html.haml
