+++ app/views/products/index.html.haml
@@ -1,4 +1,4 @@
-%h1 Product#show
+%h1 Index Product
 %ul
   - @products.each do |product|
     %li= link_to product.title, product_path(product)
