+++ app/views/products/new.html.haml
@@ -1,4 +1,4 @@
-%h1 Product#new
+%h3 Insert a Product
 = form_with model: @product do |form|
   = form.text_field :title, placeholder: "title"
   = form.text_field :description, placeholder: "description"
