--- test/controllers/product_controller_test.rb
@@ -1,23 +1,0 @@
-require "test_helper"
-
-class ProductControllerTest < ActionDispatch::IntegrationTest
-  test "should get new" do
-    get product_new_url
-    assert_response :success
-  end
-
-  test "should get show" do
-    get product_show_url
-    assert_response :success
-  end
-
-  test "should get edit" do
-    get product_edit_url
-    assert_response :success
-  end
-
-  test "should get delete" do
-    get product_delete_url
-    assert_response :success
-  end
-end
